Integration patterns.
Implemented channel adapter.
