+++ orb/src/main/java/com/epam/enterprise/orb/Stub.java
@@ -1,7 +1,9 @@
 package com.epam.enterprise.orb;
 
+import com.epam.enterprise.orb.chanel.Channel;
+import com.epam.enterprise.orb.chanel.ChannelAdapter;
 import com.epam.enterprise.orb.network.ClientMessage;
-import com.epam.enterprise.orb.network.Connection;
+import com.epam.enterprise.orb.chanel.SomeTextChannel;
 import com.epam.enterprise.orb.network.Param;
 
 /**
@@ -9,14 +11,13 @@ import com.epam.enterprise.orb.network.Param;
  */
 public class Stub implements Calculator {
 
-    private Connection<Double> connection = new Connection<Double>();
-
+    private Channel<Double> channel = new ChannelAdapter<Double>();
     @Override
     public Double add(Double a, Double b) {
         try {
             ClientMessage clientMessage = new ClientMessage("com.epam.enterprise.orb.Calculator", "add",
                     new Param[]{new Param(a, "java.lang.Double"), new Param(b, "java.lang.Double")});
-            return connection.execute(clientMessage);
+            return channel.execute(clientMessage);
         } catch (Exception e) {
             throw new RuntimeException(e);
         }
